+++ tethysext/workflows/controllers/workflows/__init__.py
@@ -6,8 +6,7 @@
 * Copyright: (c) Aquaveo 2018
 ********************************************************************************
 """
-# from .workflow_router import WorkflowRouter  # noqa: F401, E501
-# from .workflow_view import WorkflowView  # noqa: F401, E501
-# from .mixins import WorkflowMixin # noqa: F401, E501 
+from .workflow_router import WorkflowRouter  # noqa: F401, E501
+from .workflow_view import WorkflowView  # noqa: F401, E501
 
-# __all__ = ['WorkflowRouter', 'WorkflowView', 'WorkflowMixin']
+__all__ = ['WorkflowRouter', 'WorkflowView']
