+++ tethysext/workflows/controllers/workflows/workflow_view.py
@@ -14,7 +14,7 @@ from tethys_apps.utilities import get_active_app
 from tethys_sdk.permissions import has_permission
 from ...utilities import grammatically_correct_join # TODO CHECK THIS IMPORT
 from ...services.workflows.decorators import workflow_step_controller
-from .mixins import WorkflowViewMixin
+from ...mixins.workflow_mixins import WorkflowViewMixin
 from ..tethys_workflow_layout import TethysWorkflowLayout
 from ..utilities import get_style_for_status
 from ...models import Step
